ci: build SBCL from source.
Travis and Netlify both use Ubuntu 16 as their build distro. The
version of SBCL available in Ubuntu 16 is `1.3.1` which contains
`asdf` version `3.1.5`. `launch-program` was added in `asdf`
version `3.1.7.26`. The absence of `launch-program` means
`3bmd-ext-code-blocks` will fail to build. To address that I
decided to build SBCL from source in Travis to get a more up to date
version of `asdf` (which is bundled with `sbcl`).

This turned out to be a lovely yak shaving exercise.